+++ IBM.IAM.AWS.SecurityToken/LogType.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace IBM.IAM.AWS.SecurityToken
+{
+    public enum LogType : byte
+    {
+        Debug = 1,
+        Info = 2,
+        Warning = 3,
+        Error = 4
+    }
+}