+++ IBM.IAM.AWS.SecurityToken/Properties/AssemblyInfo.cs
@@ -5,14 +5,14 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("IBM - Identity and Access Management AWS Integration")]
-[assembly: AssemblyDescription("AWS SAML integration with IBM's Identity and Access Management")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("IBM - Identity and Access Management SAML")]
-[assembly: AssemblyCopyright("")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
+//[assembly: AssemblyTitle("IBM - Identity and Access Management AWS Integration")]
+//[assembly: AssemblyDescription("AWS SAML integration with IBM's Identity and Access Management")]
+//[assembly: AssemblyConfiguration("")]
+//[assembly: AssemblyCompany("")]
+//[assembly: AssemblyProduct("IBM - Identity and Access Management SAML")]
+//[assembly: AssemblyCopyright("")]
+//[assembly: AssemblyTrademark("")]
+//[assembly: AssemblyCulture("")]
 
 // Setting ComVisible to false makes the types in this assembly not visible 
 // to COM components.  If you need to access a type in this assembly from 
@@ -32,7 +32,7 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.1903.13")]
-[assembly: AssemblyFileVersion("1.0.1903.13")]
+//[assembly: AssemblyVersion("1.0.1905.36")]
+//[assembly: AssemblyFileVersion("1.0.1905.36")]
 
-[assembly: AssemblyInformationalVersion("1.0.90")]
+//[assembly: AssemblyInformationalVersion("1.0.113")]