+++ IBM.IAM.AWS.SecurityToken/SAML/IBMSAMLAuthenticationResponseParser.cs
@@ -9,8 +9,6 @@ namespace IBM.IAM.AWS.SecurityToken.SAML
 {
     internal class IBMSAMLAuthenticationResponseParser : IAuthenticationResponseParser
     {
-        public static Regex SAMLResponseField = new Regex("SAMLResponse\\W+value\\=\\\"([^\\\"]+)\\\"");
-
         //public SAMLAssertion Parse(string authenticationResponse)
         //{
         //    string assertion = string.Empty;
@@ -32,19 +30,11 @@ namespace IBM.IAM.AWS.SecurityToken.SAML
 
         Amazon.SecurityToken.SAML.SAMLAssertion IAuthenticationResponseParser.Parse(string authenticationResponse)
         {
-            string assertion = string.Empty;
-            MatchCollection resposne = IBMSAMLAuthenticationResponseParser.SAMLResponseField.Matches(authenticationResponse);
-            if (resposne.Count == 0)
-                throw new Amazon.Runtime.FederatedAuthenticationFailureException("Invalid credentials or an error occurred on server. No SAML response found from server's response.");
-            foreach (Match data in resposne)
-            {
-                assertion = data.Groups[1].Value;
-            }
 
             // Originally tried to use AWS SAMLAssertion, witch they set to internal for some lovely reason.
             BindingFlags flags = BindingFlags.NonPublic | BindingFlags.Instance;
             CultureInfo culture = null; // use InvariantCulture or other if you prefer
-            Amazon.SecurityToken.SAML.SAMLAssertion samlAssertion = Activator.CreateInstance(typeof(Amazon.SecurityToken.SAML.SAMLAssertion), flags, null, new object[] { assertion }, culture) as Amazon.SecurityToken.SAML.SAMLAssertion;
+            Amazon.SecurityToken.SAML.SAMLAssertion samlAssertion = Activator.CreateInstance(typeof(Amazon.SecurityToken.SAML.SAMLAssertion), flags, null, new object[] { authenticationResponse }, culture) as Amazon.SecurityToken.SAML.SAMLAssertion;
             return samlAssertion;
             //return new Amazon.SecurityToken.SAML.SAMLAssertion(assertion);
         }