+++ IBM.IAM.AWS.SecurityToken/SAML/StoredInfo.cs
@@ -13,8 +13,10 @@ namespace IBM.IAM.AWS.SecurityToken.SAML
     public class StoredInfo
     {
         public string StoreAs { get; internal set; }
-        public DateTime SessionExpires { get; internal set; }
+        public DateTime AssertionExpires { get; internal set; }
         public AmazonResourceName PrincipalArn { get; internal set; }
         public AmazonResourceName RoleArn { get; internal set; }
+        public DateTime Expires { get; internal set; }
+        public string AssertionDoc { get; internal set; }
     }
 }